refactor(benchmarks): organize by primitive and operation
